Add harn-serve ACP adapter
