+++ crates/harn-serve/src/adapters/acp/builtins.rs
