+++ crates/harn-serve/src/adapters/mod.rs
@@ -1,2 +1,3 @@
 pub mod a2a;
+pub mod acp;
 pub mod mcp;
