+++ docs/src/cli-reference.md
@@ -504,6 +504,9 @@ harn acp                    # bridge mode, no pipeline
 harn acp pipeline.harn      # execute a pipeline per prompt
 ```
 
+This is a legacy compatibility entrypoint. Prefer `harn serve acp <file.harn>`
+for new ACP host configuration.
+
 See [MCP and ACP Integration](./mcp-and-acp.md) for protocol details.
 
 ## harn portal
@@ -792,6 +795,7 @@ Start a workflow server through one of the outbound transport adapters.
 harn serve a2a agent.harn                  # explicit A2A
 harn serve agent.harn                      # legacy A2A shorthand
 harn serve --port 3000 agent.harn          # legacy A2A shorthand with custom port
+harn serve acp agent.harn                  # ACP session server over stdio
 harn serve mcp server.harn                 # exported pub fn -> MCP tools over stdio
 harn serve mcp --transport http server.harn
 ```
@@ -808,6 +812,12 @@ agent card at `/.well-known/a2a-agent` and supports task send, send-and-wait,
 streaming/resubscribe, push callback registration, and cancel propagation. The
 legacy shorthand `harn serve <file>` is preserved and rewrites internally to
 `harn serve a2a <file>`.
+
+`harn serve acp` starts the packaged ACP adapter on stdio for editor and IDE
+hosts. It creates ACP sessions, executes the target pipeline for each
+`session/prompt`, streams `AgentEvent` values as `session/update`
+notifications, and forwards permission prompts through
+`session/request_permission`.
 
 See [MCP and ACP Integration](./mcp-and-acp.md) and
 [Outbound workflow server](./harn-serve.md) for protocol details.
